requirements.txt
streamlit
supabase
python-dotenv
requests
plotly
pandas
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
google-auth-httplib2
